extracted Entry<T> interface; renamed previous Entry abstract class to BaseEntry; ripple
